add twopointer
